requirements.txt
aiogram==2.23.1
aiohttp==3.8.3
aiosignal==1.3.1
altgraph==0.17.3
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.1.0
Babel==2.9.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
cryptography==38.0.4
distlib==0.3.6
Django==4.1.2
django-cleanup==6.0.0
django-selenium==0.8
djangorestframework==3.14.0
environs==8.0.0
filelock==3.8.2
frozenlist==1.3.3
future==0.18.2
generics==6.0.0
h11==0.14.0
idna==3.4
magic-filter==1.0.9
marshmallow==3.19.0
multidict==6.0.3
mysql-connector-python==8.0.31
mysqlclient==2.1.1
outcome==1.2.0
packaging==22.0
pefile==2022.5.30
platformdirs==2.6.0
protobuf==3.20.1
psycopg2==2.9.5
pyarmor==7.7.4
pycparser==2.21
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
PySocks==1.7.1
python-dotenv==0.21.0
pytz==2022.6
pywin32-ctypes==0.2.0
requests==2.28.1
selenium==4.5.0
Slowloris==0.2.5
sniffio==1.3.0
sortedcontainers==2.4.0
speedtest-cli==2.1.3
sqlparse==0.4.3
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.5
urllib3==1.26.12
virtualenv==20.17.1
virtualenvwrapper-win==1.2.7
webdriver-manager==3.8.5
wsproto==1.2.0
yarl==1.8.2
